+++ final_sensors/soil_moisture.h
@@ -4,6 +4,6 @@
 #include "mcp3008.h"
 
 void soil_moisture_init( unsigned int channel );
-int soil_moisture_read(void); // Read soil moisture AOUT through MCP3008 Channel 0
+float soil_moisture_read(void); // Read soil moisture AOUT through MCP3008 Channel 0
 
 #endif
